move gui related parts in own source folder "src/gui"
Also to have more clean and files better sorted.
