+++ src/gui/channeladd.cpp
