+++ src/gui/channelsettings.cpp
