+++ src/gui/channelsettings.h
