+++ src/gui/renderingcontrol.h
